Add Dockerfile and rename src/ -> server/
