+++ README.md
@@ -4,14 +4,9 @@ Server that helps the Box to announce its local IP address without relying on mD
 ## Usage
 Run the server:
 ````bash
-npm install
-sudo node src/index.js
+docker build -t dns-server .
+docker run -d --net=host dns-server
 ````
-
-Next steps:
-* Retrieve records from disk on startup, and start with root records
-* You can only update <hash>.useraddress.net
-* Add a secret to QR code URL path
 
 Add a few records:
 ````bash
@@ -21,6 +16,8 @@ curl -i -X POST -d "{\"type\":\"A\",\"value\":\"192.168.0.42\"}" \
   http://localhost:5300/v1/dns/org/asdf/bla/192-168-0-42
 ````
 
+NB: On MacOS, Docker runs inside a virtual machine, probably on 192.168.99.100, so you may need to use that instead of 'localhost'.
+
 Query them:
 ````bash
 dig A 192-168-0-42.bla.asdf.org @localhost
@@ -29,7 +26,11 @@ dig TXT _acme-challenge.192-168-0-42.bla.asdf.org @localhost
 
 Run the tests:
 ````bash
-jshint src/
-jscs src/
+cd server
+npm install
+jshint *.js
+jscs *.js
+mv node_modules ..
+cd ..
 sudo node test/dns-query_test.js
 ````
